grok-admaster/server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pydantic-settings>=2.0.0
langchain-mcp-adapters
fastmcp
langchain
langchain-openai
mcp
deepagents
tavily-python
google-search-results
pytest>=7.0.0
pytest-asyncio>=0.21.0
scipy>=1.10.0
aiohttp>=3.9.0
httpx>=0.25.0
scikit-learn>=1.3.0
numpy>=1.24.0
torch>=2.0.0
sentence-transformers>=2.2.2
pgvector>=0.2.0
redis>=5.0.0
psutil>=5.9.0
